refactor(analyze): extract cache.ts and discovery.ts (#6)
- Create tools/analyze/cache.ts with computeFileHash, loadCache, saveCache
- Create tools/analyze/discovery.ts with IGNORE_DIRS, IGNORE_FILES, discoverFiles, getLanguageProfile
- Update index.ts to import from these modules instead of inlining the logic
- Remove now-unused imports (createHash, readdirSync, statSync, writeFileSync, extname, getStateFilePath)
